% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMCell_setup.R
\name{adapted_TM}
\alias{adapted_TM}
\title{Template matching, adapted}
\usage{
adapted_TM(
  expDat,
  sampTab,
  pThresh = 0.05,
  cThresh = 0.1,
  annotation_column = "celltype"
)
}
\arguments{
\item{expDat}{expression data}

\item{sampTab}{sample table}

\item{pThresh}{}

\item{annotation_column}{}
}
\description{
Template matching, adapted
}
