% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMCell_setup.R
\name{cn_make_tVals}
\alias{cn_make_tVals}
\title{From CellNet: Estimate gene expression dist in CTs}
\usage{
cn_make_tVals(expDat, sampTab, dLevel = "description1", predictSD = FALSE)
}
\arguments{
\item{expDat}{training data}

\item{sampTab, }{\subsection{training sample table}{
}}

\item{dLevel="description1", }{\subsection{column to define CTs}{
}}

\item{predictSD=FALSE}{\subsection{whether to predict SD based on expression level}{
}}
}
\description{
Calculate mean and SD
}
