% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMCell_setup.R
\name{compute_node_weights}
\alias{compute_node_weights}
\title{Compute node weights based on expression}
\usage{
compute_node_weights(
  nodes,
  expNorm,
  sampTab,
  source,
  target,
  annotation_column = "cf_annotation",
  method = "weighted_expression_difference",
  min_node_weight = 0.01
)
}
\arguments{
\item{expNorm}{Normalized and transformed expression data}

\item{sampTab}{Metadata/phenodata of the samples}

\item{source}{Starting cell type or background for which to compute DE}

\item{target}{Target cell type}

\item{annotation_column}{column in sampTab with cell identity annotation}

\item{method}{either target_mean_expression or weighted_expression_difference}

\item{min_node_weight}{the minimum node weight}
}
\description{
Compute node weights based on expression
}
