% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{downsample_and_transform}
\alias{downsample_and_transform}
\title{Downsamples and log-transform counts data}
\usage{
downsample_and_transform(
  counts,
  xFact = 1e+05,
  observations_by_features = FALSE
)
}
\arguments{
\item{counts}{raw counts expression matrix}

\item{xFact}{what to dowmsample and normalize to}

\item{observations_by_features}{row and column orientation of the counts matrix}
}
\value{
normalized and transformed expression matrix
}
\description{
Downsamples and log-transform counts data
}
