% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMCell_setup.R
\name{find_differential_nodes}
\alias{find_differential_nodes}
\title{Performs DE via negative binomial method on TFs using DESeq2}
\usage{
find_differential_nodes(
  expMat,
  sampTab,
  source,
  target,
  annotation_column,
  limit_to = NULL,
  pThresh = 0.05,
  lfcThresh = 1,
  topX = NULL,
  min_de_targets = 5
)
}
\arguments{
\item{expMat}{Genes-by-samples expression matrix}

\item{sampTab}{Metadata/phenodata of the samples}

\item{source}{Starting cell type or background for which to compute DE}

\item{target}{Target cell type}

\item{limit_to}{limit analysis to a subset of genes, e.g. TFs}

\item{pThresh}{adjusted p-value threshold}

\item{lfcThresh}{log-fold-change threshold}

\item{topX}{limit to_activate and to_repress to the top X genes based on LFC}

\item{min_de_targets}{if DESeq2 returns fewer targets than this threshold, T-test will be run instead}
}
\value{
A list of genes to activate and to repress
}
\description{
Performs DE via negative binomial method on TFs using DESeq2
}
