% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise_IMCell.R
\name{find_dyn_targets}
\alias{find_dyn_targets}
\title{Find targets to activate and repress in dynamic trajectories at intermediate and final states}
\usage{
find_dyn_targets(dyntraj, state_traj = NULL, column_annotation = "celltype")
}
\arguments{
\item{dyntraj}{the result of running static_to_dynamic_trajectory}

\item{state_traj}{starting, intermediate, and final cell states}

\item{column_annotation}{column name in sampTab within dyntraj specifying cell states}
}
\value{

}
\description{
Find targets to activate and repress in dynamic trajectories at intermediate and final states
}
