% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IM.R
\name{im_celf_parallel}
\alias{im_celf_parallel}
\title{Cost Effective Lazy Forward algorithm (CELF)}
\usage{
im_celf_parallel(
  ig,
  kmax = 5,
  tfs = NULL,
  targets = NULL,
  niter = 1000,
  edge_probability_method = "in_degree",
  edge_probability_multiplier = 1,
  min_marginal_spread = 2,
  return_spread = FALSE,
  num_cores = 2
)
}
\arguments{
\item{ig}{an igraph object with normalized edge attribute 'weight' corresponding to probabilities}

\item{targets}{consider activation of specified targets only. If NULL activation is based on total spread}

\item{min_marginal_spread}{the minimum marginal spread. Greedy algorithm stops when marginal spread < minimum marginal spread. If minimum_marginal_spread=1, algorithm doesn't terminate until k=kmax or spread reaches full coverage}

\item{return_spread}{whether or not to return the set of activated nodes}

\item{seed}{starting nodes to activate}
}
\value{

}
\description{
Cost Effective Lazy Forward algorithm (CELF)
}
