% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMCell.R
\name{im_combined_PRIM_greedy_testing}
\alias{im_combined_PRIM_greedy_testing}
\title{Greedy Algorithm for combined PRIM (spread = targets activated + targets repressed)}
\usage{
im_combined_PRIM_greedy_testing(
  ig,
  kmax = 5,
  repressor_wins = FALSE,
  tfs = NULL,
  targets_activate = NULL,
  targets_repress = NULL,
  niter = 1000,
  edge_probability_method = "in_degree",
  edge_probability_multiplier = 1,
  min_marginal_spread = 5,
  return_spread = FALSE,
  num_cores = 2,
  temptesting = TRUE
)
}
\arguments{
\item{ig}{an igraph object with normalized edge attribute 'weight' corresponding to probabilities, 'type' corresponding to interaction type,}

\item{kmax}{max number of TFs in solution set}

\item{repressor_wins}{See random_ICP}

\item{tfs}{limit search radius to specific TFs. If NULL, search radius is not limited.}

\item{targets_activate}{consider activation of specified targets only. If NULL activation is based on total spread}

\item{targets_repress}{consider repression of specified targets only. If NULL activation is based on total spread}

\item{niter}{niter in MC}

\item{edge_probability_method}{method to assign edge probabilities}

\item{min_marginal_spread}{the minimum marginal spread. Greedy algorithm stops when marginal spread < minimum marginal spread. If minimum_marginal_spread=1, algorithm doesn't terminate until k=kmax or spread reaches full coverage}

\item{return_spread}{whether or not to return the set of activated nodes}

\item{num_cores}{number of cores}
}
\description{
Greedy Algorithm for combined PRIM (spread = targets activated + targets repressed)
}
