% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMCell.R
\name{im_single_PRIM_celf_parallel}
\alias{im_single_PRIM_celf_parallel}
\title{Cost Effective Lazy Forward algorithm (CELF) for single signed IM (either PIM or NIM) in PARALLEL}
\usage{
im_single_PRIM_celf_parallel(
  ig,
  kmax = 5,
  mode = "PIM",
  tfs = NULL,
  repressor_wins = FALSE,
  targets = NULL,
  niter = 1000,
  edge_probability_method = "in_degree",
  edge_probability_multiplier = 1,
  min_marginal_spread = 5,
  return_spread = FALSE,
  num_cores = 2
)
}
\arguments{
\item{ig}{an igraph object with normalized edge attribute 'weight' corresponding to probabilities}

\item{kmax}{max number of TFs in solution set}

\item{mode}{either "PIM" or "NIM" -- maximize activation or repression}

\item{tfs}{limit search radius to specific TFs. If NULL, search radius is not limited.}

\item{repressor_wins}{See random_ICP}

\item{targets}{consider activation (or repression is mode="NIM") of specified targets only. If NULL activation is based on total spread}

\item{edge_probability_method}{method to assign edge probabilities}

\item{edge_probability_multiplier}{optional factor to multiply edge probabilities by. 1 = no multiplier}

\item{min_marginal_spread}{the minimum marginal spread. Greedy algorithm stops when marginal spread < minimum marginal spread. If minimum_marginal_spread=1, algorithm doesn't terminate until k=kmax or spread reaches full coverage}

\item{return_spread}{whether or not to return the set of influenced nodes}
}
\value{

}
\description{
Cost Effective Lazy Forward algorithm (CELF) for single signed IM (either PIM or NIM) in PARALLEL
}
