% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMCell.R
\name{IMCell_CELF}
\alias{IMCell_CELF}
\title{Cost Effective Lazy Forward algorithm (CELF)}
\usage{
IMCell_CELF(
  ig,
  kmax = 5,
  tfs = NULL,
  targets = NULL,
  niter = 1000,
  edge_probability_method = "in_degree",
  edge_probability_multiplier = 1,
  min_marginal_spread = 5,
  return_spread = FALSE,
  num_cores = 2
)
}
\arguments{
\item{ig}{an igraph object with normalized edge attribute 'weight' corresponding to probabilities}

\item{kmax}{maximum number of TFs in the solution set}

\item{tfs}{TFlist}

\item{targets}{consider activation of specified targets only. If NULL activation is based on total spread}

\item{niter}{number of random cascades to run}

\item{edge_probability_method}{method to convert edge weights to probabilities}

\item{edge_probability_multiplier}{optional factor to multiply edge probabilities by. 1 = no multiplier}

\item{min_marginal_spread}{the minimum marginal spread. Greedy algorithm stops when marginal spread < minimum marginal spread. If minimum_marginal_spread=1, algorithm doesn't terminate until k=kmax or spread reaches full coverage}

\item{return_spread}{whether or not to return the set of activated nodes}

\item{num_cores}{number of cores to use if running parallel}
}
\value{

}
\description{
Cost Effective Lazy Forward algorithm (CELF)
}
