% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise_IMCell.R
\name{IMCell_epochnets_expweighted}
\alias{IMCell_epochnets_expweighted}
\title{Run IMCell_expweighted successively on dynamic network}
\usage{
IMCell_epochnets_expweighted(
  dynamic_grnDF,
  network_path,
  targets = NULL,
  expX = NULL,
  sampTab = NULL,
  epochs = NULL,
  tfs = NULL,
  kmax = 5,
  niter = 1000,
  min_marginal_spread = 5,
  repressor_wins = FALSE,
  node_weight_method = "no_weight",
  min_node_weight = 0.01,
  edge_probability_method = "in_degree",
  edge_probability_multiplier = 1,
  num_cores = 2,
  return_spread = FALSE,
  temptesting = TRUE
)
}
\arguments{
\item{dynamic_grnDF}{a dynamic GRN in dataframe form}

\item{network_path}{the names and order of the subnetworks in dynamic_grnDF}

\item{sampTab}{meta data on cells, including pseudotime}

\item{epochs}{epoch names. If NULL inferred from network_path}

\item{tfs}{limit search radius to specific TFs. If NULL, search radius is not limited.}

\item{kmax}{max number of TFs in solution set}

\item{niter}{niter in MC}

\item{min_marginal_spread}{the minimum marginal spread. Greedy algorithm stops when marginal spread < minimum marginal spread. If minimum_marginal_spread=1, algorithm doesn't terminate until k=kmax or spread reaches full coverage}

\item{repressor_wins}{See random_ICP}

\item{node_weight_method}{node weighitng method}

\item{min_node_weight}{minimum node weight}

\item{edge_probability_method}{method to assign edge probabilities}

\item{edge_probability_multiplier}{optional factor to multiply edge probabilities by. 1 = no multiplier}

\item{num_cores}{number of cores}

\item{return_spread}{whether or not to return the set of activated nodes}
}
\value{

}
\description{
Run IMCell_expweighted successively on dynamic network
}
