% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMCell.R
\name{IMCell_expweighted}
\alias{IMCell_expweighted}
\title{Greedy MC Algorithm for IMCell, with optional expression weighting}
\usage{
IMCell_expweighted(
  ig,
  expDat,
  sampTab,
  source,
  target,
  annotation_column = "annotation_column",
  node_weight_method = "weighted_expression_difference",
  min_node_weight = 0.01,
  kmax = 5,
  repressor_wins = FALSE,
  tfs = NULL,
  targets_activate = NULL,
  targets_repress = NULL,
  niter = 1000,
  edge_probability_method = "in_degree",
  edge_probability_multiplier = 1,
  min_marginal_spread = 10,
  return_spread = FALSE,
  num_cores = 2,
  temptesting = FALSE
)
}
\arguments{
\item{ig}{an igraph object with normalized edge attribute 'weight' corresponding to probabilities, 'type' corresponding to interaction type}

\item{expDat}{expression data}

\item{sampTab}{sample table}

\item{source}{the starting cell type}

\item{target}{the target cell type}

\item{annotation_column}{the column in sampTab with cell identity annotation}

\item{node_weight_method}{method to weight the nodes}

\item{min_node_weight}{minimum assigned node weight}

\item{kmax}{max number of TFs in solution set}

\item{repressor_wins}{See random_ICP}

\item{tfs}{limit search radius to specific TFs. If NULL, search radius is not limited.}

\item{targets_activate}{consider activation of specified targets only. If NULL activation is based on total spread}

\item{targets_repress}{consider repression of specified targets only. If NULL activation is based on total spread}

\item{niter}{niter in MC}

\item{edge_probability_method}{method to assign edge probabilities}

\item{edge_probability_multiplier}{optional factor to multiply edge probabilities by. 1 = no multiplier}

\item{min_marginal_spread}{the minimum marginal spread. Greedy algorithm stops when marginal spread < minimum marginal spread. If minimum_marginal_spread=1, algorithm doesn't terminate until k=kmax or spread reaches full coverage}

\item{return_spread}{whether or not to return the set of activated nodes}

\item{num_cores}{number of cores}
}
\value{

}
\description{
Greedy MC Algorithm for IMCell, with optional expression weighting
}
