% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IM.R, R/IMCell.R
\name{mc_activation_cascade_parallel}
\alias{mc_activation_cascade_parallel}
\title{run MC cascade and return expected spread, assumes edge weights are positive}
\usage{
mc_activation_cascade_parallel(
  ig,
  seed,
  targets = NULL,
  niter = 1000,
  return_spread = FALSE,
  num_cores = 2
)

mc_activation_cascade_parallel(
  ig,
  seed,
  targets = NULL,
  niter = 1000,
  return_spread = FALSE,
  num_cores = 2
)
}
\arguments{
\item{ig}{an igraph object with normalized edge attribute 'weight' corresponding to probabilities}

\item{seed}{starting nodes to activate}

\item{targets}{consider activation of specified targets only. If NULL activation is based on total spread}

\item{return_spread}{whether or not to return the set of activated nodes}

\item{num_cores}{number of cores to use if running parallel}
}
\value{

}
\description{
run MC cascade and return expected spread, assumes edge weights are positive

run MC cascade and return expected spread, assumes edge weights are positive
}
