% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMCell.R
\name{mc_ICP_expweighted}
\alias{mc_ICP_expweighted}
\title{run MC cascade and return expected spread on IC-P model}
\usage{
mc_ICP_expweighted(
  ig,
  seed,
  node_weights_activate,
  node_weights_repress,
  repressor_wins = FALSE,
  targets_activate = NULL,
  targets_repress = NULL,
  niter = 1000,
  return_spread = FALSE,
  num_cores = 2
)
}
\arguments{
\item{ig}{an igraph object with normalized edge attribute 'weight' corresponding to probabilities, and 'type' corresponding to interaction type}

\item{seed}{starting nodes to activate}

\item{repressor_wins}{see random_ICP for description}

\item{targets_activate}{consider activation of specified targets only. If NULL activation is based on total spread}

\item{targets_repress}{consider repression of specified targets only. If NULL activation is based on total spread}

\item{return_spread}{whether or not to return the set of activated nodes}
}
\value{

}
\description{
run MC cascade and return expected spread on IC-P model
}
