% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IM.R, R/IMCell.R
\name{random_activation_cascade_parallel}
\alias{random_activation_cascade_parallel}
\title{Simulates a random cascade (activation only)}
\usage{
random_activation_cascade_parallel(
  ig,
  seed,
  targets = NULL,
  return_spread = FALSE
)

random_activation_cascade_parallel(
  ig,
  seed,
  targets = NULL,
  return_spread = FALSE
)
}
\arguments{
\item{ig}{an igraph object with normalized edge attribute 'weight' corresponding to probabilities}

\item{seed}{starting nodes to activate}

\item{targets}{consider activation of specified targets only. If NULL activation is based on total spread}

\item{return_spread}{whether or not to return the set of activated nodes}
}
\value{

}
\description{
Simulates a random cascade (activation only)

Simulates a random cascade (activation only)
}
