% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMCell.R
\name{rank_by_spread}
\alias{rank_by_spread}
\title{Compute marginal spread with combined PRIM rules (spread = targets activated + targets repressed)}
\usage{
rank_by_spread(
  ig,
  repressor_wins = FALSE,
  tfs = NULL,
  targets_activate = NULL,
  targets_repress = NULL,
  niter = 1000,
  edge_probability_method = "in_degree",
  edge_probability_multiplier = 1,
  return_spread = FALSE,
  num_cores = 2
)
}
\arguments{
\item{ig}{an igraph object with normalized edge attribute 'weight' corresponding to probabilities, 'type' corresponding to interaction type,}

\item{repressor_wins}{See random_ICP}

\item{tfs}{limit search radius to specific TFs. If NULL, search radius is not limited.}

\item{targets_activate}{consider activation of specified targets only. If NULL activation is based on total spread}

\item{targets_repress}{consider repression of specified targets only. If NULL activation is based on total spread}

\item{niter}{niter in MC}

\item{edge_probability_method}{method to assign edge probabilities}

\item{return_spread}{whether or not to return the set of activated nodes}

\item{num_cores}{number of cores}

\item{kmax}{max number of TFs in solution set}
}
\value{

}
\description{
Compute marginal spread with combined PRIM rules (spread = targets activated + targets repressed)
}
