% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise_IMCell.R
\name{static_to_dynamic_trajectory}
\alias{static_to_dynamic_trajectory}
\title{Convert static to dynamic network}
\usage{
static_to_dynamic_trajectory(
  grnDF,
  expX,
  sampTab,
  pseudotime_column = "pseudotime",
  epoch_annotation_column = NULL,
  num_epochs = 2,
  path = NULL,
  column_annotation = NULL
)
}
\arguments{
\item{grnDF}{a static GRN in dataframe form}

\item{sampTab}{meta data on cells, including pseudotime}

\item{pseudotime_column}{column name in sampTab with pseudotime annotation}

\item{epoch_annotation_column}{column name in sampTab that has pre-assigned epochs. If not NULL, num_epochs is ignored. Otherwise, epochs are computed via pseudotime.}

\item{num_epochs}{number of epochs, ignored if epoch_annotation_column is not NULL.}

\item{path}{path through the trajectory}

\item{column_annotation}{}
}
\value{

}
\description{
Convert static to dynamic network
}
