% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMCell_setup.R
\name{tfscope}
\alias{tfscope}
\title{Limit TF scope and search radius}
\usage{
tfscope(
  expCounts,
  sampTab,
  tfs,
  source,
  target,
  annotation_column = "celltype",
  sourceThresh = NULL,
  DE_lfcThresh = 1,
  DE_pThresh = 0.05,
  DE_pval_topX = 20,
  DE_lfc_topX = 20,
  TM_pThresh = 0.05,
  TM_cThresh = 0.1,
  TM_topX = 20,
  TM_topX_sourceThresh = 50
)
}
\arguments{
\item{expCounts}{Raw counts expression data}

\item{sampTab}{Metadata/phenodata of the samples}

\item{tfs}{list of TFs}

\item{source}{Starting cell type or background for which to compute DE}

\item{target}{Target cell type}

\item{annotation_column}{column in sampTab with cell identity annotation}

\item{sourceThresh}{Filter out TFs if they are expressed above this level in the source cell type}

\item{DE_lfcThresh}{Log-fold-change threshold for DE approach}

\item{DE_pThresh}{adjusted p-value threshold for DE approach}

\item{DE_pval_topX}{number of top TFs to keep by DE p-value}

\item{DE_lfc_topX}{number of top TFs to keep by DE LFC}

\item{TM_pThresh}{adjusted p-value threshold for TM approach}

\item{TM_cThresh}{corr threshold for TM approach}

\item{TM_topX}{number of top TFs to keep by TM appraoch}

\item{TM_topX_sourceThresh}{Filter out if considered spec gene in source type}
}
\value{
narrowed TF search radius
}
\description{
Limit TF scope and search radius
}
