% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMCell_setup.R
\name{trans_rnaseq}
\alias{trans_rnaseq}
\title{weighted subtraction from mapped reades and log applied to all}
\usage{
trans_rnaseq(expRaw, total)
}
\arguments{
\item{expRaw}{matrix of total mapped reads per gene/transcript}

\item{total}{numeric post transformation sum of read counts}
}
\description{
Simulate expression profile of  \emph{total} mapped reads
}
