% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMCell_setup.R
\name{weights_to_probability}
\alias{weights_to_probability}
\title{Converts edge weights to probabilities}
\usage{
weights_to_probability(
  ig,
  method = "in_degree",
  multiplier = 1,
  p = 0.5,
  pmax = 0.8
)
}
\arguments{
\item{ig}{a directed igraph object}

\item{method}{how to convert weights to probability. Either in_degree, total_in_degree, or coin_flip.}

\item{multiplier}{a multiplier to apply to probabilities. Ignored if method is coin_flip.}

\item{p}{probability if method is coin_flip, ignored otherwise.}

\item{pmax}{value at which to cap the probability.}
}
\value{
updated igraph object with weight now updated as probabilities
}
\description{
Converts edge weights to probabilities
}
